requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py311h6c40b1e_1002
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.08.22=hecd8cb5_0
certifi=2023.7.22=py311hecd8cb5_0
cffi=1.15.1=py311h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.5=py311ha357a0b_0
cryptography=41.0.3=py311h30e54ef_0
cycler=0.11.0=pyhd3eb1b0_0
filelock=3.9.0=py311hecd8cb5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py311h1c2e9e1_0
idna=3.4=py311hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43547
jinja2=3.1.2=py311hecd8cb5_0
jpeg=9e=h6c40b1e_1
kiwisolver=1.4.4=py311hcec6c5f_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libtiff=4.5.1=hcec6c5f_0
libuv=1.44.2=h6c40b1e_0
libwebp=1.3.2=hf6ce154_0
libwebp-base=1.3.2=h6c40b1e_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.1=py311h6c40b1e_0
matplotlib=3.7.2=py311hecd8cb5_0
matplotlib-base=3.7.2=py311h8251f7d_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.8=py311h6c40b1e_0
mkl_random=1.2.4=py311ha357a0b_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py311hecd8cb5_0
munkres=1.1.4=py_0
ncurses=6.4=hcec6c5f_0
networkx=3.1=py311hecd8cb5_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
numpy=1.26.0=py311h728a8a3_0
numpy-base=1.26.0=py311h53bf9ac_0
openssl=3.0.11=hca72f7f_2
packaging=23.1=py311hecd8cb5_0
pillow=9.4.0=py311hcec6c5f_1
pip=23.2.1=py311hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311hecd8cb5_0
pyparsing=3.0.9=py311hecd8cb5_0
pysocks=1.7.1=py311hecd8cb5_0
python=3.11.5=hf27a42d_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=2.0.1=cpu_py311he1daf47_0
readline=8.2=hca72f7f_0
requests=2.31.0=py311hecd8cb5_0
setuptools=68.0.0=py311hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h6c40b1e_0
sympy=1.11.1=py311hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.12=h5d9f67b_0
torchvision=0.15.2=cpu_py311h5c52452_0
tornado=6.3.2=py311h6c40b1e_0
typing-extensions=4.7.1=py311hecd8cb5_0
typing_extensions=4.7.1=py311hecd8cb5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hecd8cb5_0
wheel=0.41.2=py311hecd8cb5_0
xz=5.4.2=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
